requirements.txt
# Lead Management System - Dependencies
fastapi==0.115.0
uvicorn==0.30.6
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
bcrypt==4.0.1
python-dotenv==1.2.1
pillow

# Phase 2: Excel upload
openpyxl==3.1.5

# Phase 3: OCR
groq==1.1.2

# Phase 4: Email + WhatsApp
aiosmtplib==3.0.2
